Move functions not used by the webhook view to the corresponding module
